+++ pykechain/__about__.py
@@ -2,7 +2,7 @@
 name = 'pykechain'
 description = 'KE-chain Python SDK'
 
-version = '1.16.0'
+version = '2.0.0-rc6'
 
 author = 'KE-works BV'
 email = 'support+pykechain@ke-works.com'