+++ tests/test_properties.py
@@ -75,7 +75,7 @@ class TestProperties(TestBetamax):
 
         # test creation of new property model of bike
         new_property = self.client.create_property(model=bike, name='New property', description='Nice prop',
-                                                   property_type='CHAR', default_value='EUREKA!')
+                                                   property_type=PropertyType.CHAR_VALUE, default_value='EUREKA!')
 
         # check whether the property has been created and whether it's name and type are correct
         self.assertIsInstance(bike.property('New property'), Property)
@@ -153,13 +153,13 @@ class TestProperties(TestBetamax):
 
     def test_edit_property_model_unit(self):
         front_fork_model = self.project.model('Front Fork')
-        height_property = front_fork_model.property(name='Height (mm)')
+        height_property = front_fork_model.property(name='Height')
         height_old_unit = str(height_property._json_data.get('unit'))
         new_unit = 'm'
 
         height_property.edit(unit=new_unit)
 
-        height_property_u = front_fork_model.property(name='Height (mm)')
+        height_property_u = front_fork_model.property(name='Height')
         self.assertEqual(height_property.id, height_property_u.id)
 
         with self.assertRaises(IllegalArgumentError):