+++ tests/utils.py
@@ -1,20 +1,23 @@
+from contextlib import contextmanager
+
 import os
 import six
 from envparse import Env
 
 # reads a local .env file with the TEST_TOKEN=<user token>
 # ensure that this file is not commited to github (never ever)
-from contextlib import contextmanager
-
 env = Env()
 env.read_envfile()
 
-TEST_URL = env('TEST_URL', default='https://kec2api.ke-chain.com')
+TEST_URL = env('TEST_URL', default='https://kec3api.ke-chain.com')
 TEST_USERNAME = env('TEST_USERNAME', default='pykechain')  # LVL1
 TEST_TOKEN = env('TEST_TOKEN', default='')
 TEST_SCOPE_ID = env('TEST_SCOPE_ID', default='6f7bc9f0-228e-4d3a-9dc0-ec5a75d73e1d')
 TEST_SCOPE_NAME = env('TEST_SCOPE_NAME', default='Bike Project (pykechain testing)')
 TEST_RECORD_CASSETTES = env.bool('TEST_RECORD_CASSETTES', default=True)
+
+# flags for altering testing behaviour (to skip test) for major API changes requiring different tests.
+TEST_FLAG_IS_WIM2 = env.bool('TEST_FLAG_IS_WIM2', default=True)
 
 
 @contextmanager
@@ -32,7 +35,7 @@ def temp_chdir(cwd=None):
     else:
         from tempfile import mkdtemp
         tempwd = mkdtemp()
-        origin=cwd or os.getcwd()
+        origin = cwd or os.getcwd()
         os.chdir(tempwd)
         try:
             yield tempwd if os.path.exists(tempwd) else ''